Large conversion from size_t to SDVN or uint64_t

+++ ibrdtn/daemon/src/api/BinaryStreamClient.cpp
@@ -124,7 +124,7 @@ namespace dtn
 			}
 		}
 
-		void BinaryStreamClient::eventBundleAck(size_t ack) throw ()
+		void BinaryStreamClient::eventBundleAck(uint64_t ack) throw ()
 		{
 			_lastack = ack;
 		}
