+++ ibrdtn/daemon/src/core/BundleCore.h
@@ -114,8 +114,8 @@ namespace dtn
 			void raiseEvent(const dtn::core::Event *evt) throw ();
 
 			virtual void validate(const dtn::data::PrimaryBlock &obj) const throw (RejectedException);
-			virtual void validate(const dtn::data::Block &obj, const size_t length) const throw (RejectedException);
-			virtual void validate(const dtn::data::PrimaryBlock &bundle, const dtn::data::Block &obj, const size_t length) const throw (RejectedException);
+			virtual void validate(const dtn::data::Block &obj, const uint64_t length) const throw (RejectedException);
+			virtual void validate(const dtn::data::PrimaryBlock &bundle, const dtn::data::Block &obj, const uint64_t length) const throw (RejectedException);
 			virtual void validate(const dtn::data::Bundle &obj) const throw (RejectedException);
 			virtual void validate(const dtn::data::MetaBundle &obj) const throw (RejectedException);
